+++ src/config.js
@@ -3,8 +3,6 @@
 
 	$.fn.loading.defaults = {
 		opacity:    0.9,
-		//speedIn:    300,
-		//speedOut:   1000,
 
 		algorithm: {
 			name: 'snake',
@@ -12,23 +10,28 @@
 				reverse: false
 			}
 		},
-		effect:    ['simple'],
+
+		effect: ['simple'],
 
 		spinner: {
-			width:   35,
-			height:  35,
+			width:   "50%",
+			height:  "30%",
 			matrix: {
 				x: null,
 				y: null
 			},
 			pin: {
-				width:  7,
-				height: 7,
+				width:  5,
+				height: 5,
 				margin: {
 					top:    1,
 					right:  1,
 					bottom: 0,
 					left:   0
+				},
+				css: {
+					background: "green",
+					opacity: 0
 				}
 			},
 			interval: 100
@@ -41,7 +44,8 @@
 			// raw css
 			css: {
 				opacity: 0.8,
-				"z-index": 9999
+				"z-index": 9999,
+				overflow: "hidden"
 			}
 		}
 	};