+++ src/effect/fancy.js
@@ -1,6 +1,10 @@
 (function($) {
     "use strict";
 
+	var defaults = {
+
+	};
+
     $.fn.loading.effect('fancy', function(pin, interval, runtime) {
         pin
             .css({
@@ -8,5 +12,5 @@
             })
             .animate({width: '-=2', height: '-=2', top: '+=1', left: '+=1'}, interval/2)
             .animate({width: '+=2', height: '+=2', top: '-=1', left: '-=1'}, interval/2);
-    });
+    }, defaults);
 }).call(this, jQuery);