+++ src/effect/simple-progress.js
@@ -1,6 +1,10 @@
 (function($) {
     "use strict";
 
+	var defaults = {
+
+	};
+
     $.fn.loading.effect('simple-progress', function(pin, interval, runtime) {
         if (!pin.data('simple-progress-init')) {
             pin
@@ -16,5 +20,5 @@
             .animate({
                 opacity: pin.data('simple-progress-sign') ? 1 : runtime.progress/100
             }, interval);
-    });
+    }, defaults);
 }).call(this, jQuery);