+++ src/effect/simple.js
@@ -1,20 +1,16 @@
 (function($) {
     "use strict";
 
-    $.fn.loading.effect('simple', function(pin, interval, runtime) {
-        if (!pin.data('simple-init')) {
-            pin
-                .css({
-                    background: 'green',
-                    opacity: 0
-                })
-                .data('simple-init', true);
-        }
+	var defaults = {
 
+	};
+
+    $.fn.loading.effect('simple', function(pin, interval, runtime) {
         pin
             .data('simple-sign', pin.data('simple-sign') ? false : true)
             .animate({
                 opacity: pin.data('simple-sign') ? 1 : 0
             }, interval);
-    });
+    }, defaults);
+
 }).call(this, jQuery);